backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.0
Flask-Migrate==4.0.4
Flask-CORS==5.0.0
pycodestyle==2.12.0
tensorflow==2.17.0
Werkzeug==3.0.4
# system level dependency to be installed on ubuntu b4 installing msqlclient:
# sudo apt-get install python3-dev default-libmysqlclient-dev build-essential pkg-config
# mysqlclient==2.2.4
# mysql-connector-python==9.0.0
Pillow==10.4.0
python-dotenv=1.0.1
